+++ playground/react-app/report.ts
@@ -1,6 +1,6 @@
 async function report(coverage: any) {
   const reponame = 'react-app'
-  await fetch(`http://localhost:8988/${reponame}/coverage/client?v=temp_git_commit_id`, {
+  await fetch(`http://localhost:3000/${reponame}/coverage/client?v=temp_git_commit_id`, {
     body: JSON.stringify(coverage),
     method: 'POST',
     headers: {