+++ playground/react-app/src/App.tsx
@@ -1,5 +1,8 @@
-import { useState } from 'react'
+import { Suspense, lazy, useState } from 'react'
+import Static from './Static'
 import './App.css'
+
+const Lazy = lazy(() => import('./Lazy'))
 
 function App() {
   const [count, setCount] = useState(0)
@@ -9,6 +12,10 @@ function App() {
       <div className='card'>
         <button onClick={() => setCount((count) => count + 1)}>count is {count}</button>
       </div>
+      <Suspense fallback={<div />}>
+        <Lazy />
+      </Suspense>
+      <Static />
     </>
   )
 }