+++ playground/react-app/src/Lazy.tsx
@@ -1,0 +1,3 @@
+export default function Lazy() {
+  return <div>this is lazy loaded</div>
+}