+++ playground/react-app/src/Static.tsx
@@ -1,0 +1,3 @@
+export default function Static() {
+  return <div>this is static</div>
+}