+++ playground/vue-app/report.ts
@@ -1,6 +1,6 @@
 function report(coverage: any) {
   const reponame = 'vue-app'
-  fetch(`http://localhost:8988/${reponame}/coverage/client`, {
+  fetch(`http://localhost:3000/${reponame}/coverage/client`, {
     body: JSON.stringify(coverage),
     method: 'POST',
     headers: {