+++ packages/istanbul-server-middleware/CHANGELOG.md
@@ -1,5 +1,11 @@
 # istanbul-server-middleware
 
+## 1.1.0
+
+### Minor Changes
+
+- support merge coverage
+
 ## 1.0.0
 
 ### Major Changes